Add documentation and some stubs (#9)
* Move FolderProcessorSettings to the models project

* Added a lot of documentation

* Removed internal docs folder for the time being

* Replacing my home-rolled retry with Polly.

* Shuffled things around a bit, started stubbing some methods, and updated packages.

* Fixed a previously introduced bug that broke setting deserialization

* Updated dependencies
