+++ FolderProcessor.sln
@@ -2,17 +2,13 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FolderProcessor", "src\FolderProcessor\FolderProcessor.csproj", "{7B959D8B-8867-4F7A-A925-3E5797AC244A}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{D1457DD6-C1FB-40BF-A633-B86F43CE7527}"
-	ProjectSection(SolutionItems) = preProject
-		Docs\Overview.md = Docs\Overview.md
-	EndProjectSection
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FolderProcessor.Host", "src\FolderProcessor.Host\FolderProcessor.Host.csproj", "{61C4D197-BED1-4561-A674-AE06E6C182EA}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D4A0CB24-C8D2-49DE-9278-FD35013DAB02}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
 		.gitignore = .gitignore
+		coverlet.runsettings = coverlet.runsettings
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FolderProcessor.Abstractions", "src\FolderProcessor.Abstractions\FolderProcessor.Abstractions.csproj", "{6771A627-C999-40FF-B7B5-2C6FB60CD733}"
