+++ src/FolderProcessor.Abstractions/FolderProcessor.Abstractions.csproj
@@ -22,7 +22,7 @@
 
     <ItemGroup>
       <PackageReference Include="JetBrains.Annotations" Version="2021.3.0" />
-      <PackageReference Include="MediatR.Contracts" Version="1.0.0" />
+      <PackageReference Include="MediatR.Contracts" Version="1.0.1" />
     </ItemGroup>
 
 </Project>
