+++ src/FolderProcessor.Extensions.Microsoft.DependencyInjection/FolderProcessor.Extensions.Microsoft.DependencyInjection.csproj
@@ -18,6 +18,7 @@
 
     <ItemGroup>
       <PackageReference Include="JetBrains.Annotations" Version="2021.3.0" />
+      <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.1" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
       <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0" />
       <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
@@ -25,6 +26,10 @@
 
     <ItemGroup>
       <ProjectReference Include="..\FolderProcessor\FolderProcessor.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <Folder Include="Models" />
     </ItemGroup>
 
 </Project>
