+++ src/FolderProcessor.Host/FolderProcessor.Host.csproj
@@ -12,10 +12,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="MediatR" Version="10.0.0" />
-        <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="10.0.0" />
+        <PackageReference Include="MediatR" Version="10.0.1" />
         <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.0" />
-        <PackageReference Include="System.IO.Abstractions" Version="16.0.8" />
     </ItemGroup>
 
     <ItemGroup>
