+++ src/FolderProcessor.Host/Program.cs
@@ -1,18 +1,19 @@
 using System.IO.Abstractions;
 using FolderProcessor.Abstractions.Monitoring.Filters;
 using FolderProcessor.Extensions.Microsoft.DependencyInjection;
+using FolderProcessor.Extensions.Microsoft.DependencyInjection.Monitoring;
 using FolderProcessor.Host;
-using FolderProcessor.Host.Models;
-using FolderProcessor.Monitoring;
+using FolderProcessor.Models.Monitoring.Configuration;
 using FolderProcessor.Monitoring.Filters;
-using MediatR;
 
 var host = Host.CreateDefaultBuilder(args)
     .ConfigureServices((host, services) =>
     {
-        services.AddMediatR(typeof(Program), typeof(StreamedFolderWatcher))
+        services
+            // Automatically sets up FolderProcessor with settings loaded from IConfiguration
             .AddFolderProcessor(host.Configuration)
-            .AddFolderWatcher(() => new PollingWatcherSettings
+            // Since we setup folder processor above, we can just add ad-hoc watchers
+            .AddFolderWatcher(() => new PollingFolderWatcherSettings
             {
                 Folder = "Data/Child",
                 Interval = TimeSpan.FromSeconds(30)
