+++ src/FolderProcessor.Models/Monitoring/Configuration/WatcherSettings.cs
@@ -1,14 +1,6 @@
 using JetBrains.Annotations;
 
-// ReSharper disable once CheckNamespace
-namespace FolderProcessor.Host.Models;
-
-[PublicAPI]
-public enum WatcherType
-{
-    Polling,
-    FileSystemWatcher
-}
+namespace FolderProcessor.Models.Monitoring.Configuration;
 
 [PublicAPI]
 public class WatcherSettings
@@ -16,7 +8,7 @@ public class WatcherSettings
     public WatcherType Type { get; set; }
     public string Folder { get; set; } = null!;
 
-    protected WatcherSettings()
+    public WatcherSettings()
     {
         Type = WatcherType.FileSystemWatcher;
     }
@@ -30,25 +22,9 @@ public class WatcherSettings
     {
         return type switch
         {
-            WatcherType.Polling => new PollingWatcherSettings(folder),
+            WatcherType.Polling => new PollingFolderWatcherSettings(folder),
             WatcherType.FileSystemWatcher => new WatcherSettings(folder),
             _ => throw new ArgumentOutOfRangeException(nameof(type), type, null)
         };
-    }
-}
-
-[PublicAPI]
-public class PollingWatcherSettings : WatcherSettings
-{
-    public TimeSpan Interval { get; set; } = TimeSpan.FromSeconds(30);
-
-    public PollingWatcherSettings()
-    {
-        Type = WatcherType.Polling;
-    }
-
-    public PollingWatcherSettings(string folder) : this()
-    {
-        Folder = folder;
     }
 }
