+++ src/FolderProcessor.Models/Monitoring/Configuration/WatcherType.cs
@@ -1,0 +1,10 @@
+using JetBrains.Annotations;
+
+namespace FolderProcessor.Models.Monitoring.Configuration;
+
+[PublicAPI]
+public enum WatcherType
+{
+    Polling,
+    FileSystemWatcher
+}
