+++ src/FolderProcessor/Common/Policies/FilePolicies.cs
@@ -1,0 +1,8 @@
+namespace FolderProcessor.Common.Policies;
+
+public class FilePolicies
+{
+    public FilePolicies()
+    {
+    }
+}
