+++ test/FolderProcessor.UnitTests/FolderProcessor.UnitTests.csproj
@@ -11,9 +11,9 @@
         <PackageReference Include="AutoFixture.AutoMoq" Version="4.17.0" />
         <PackageReference Include="AutoFixture.Xunit2" Version="4.17.0" />
         <PackageReference Include="FluentAssertions" Version="6.3.0" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.0" />
         <PackageReference Include="Moq" Version="4.16.1" />
-        <PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="16.1.1" />
+        <PackageReference Include="System.IO.Abstractions.TestingHelpers" Version="16.1.4" />
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
